This should hopefully be a full treatment of the "Toggle System Maintenance Mode" use case containing the Texual descriptions (located on the Google Doc file accessible from this repository's Wiki) and the attached System Sequence and System Diagrams.
I glanced at the Class Diagram and I'm not sure how I want to go about it at the moment.
Will discuss how verbose we want to be about it in the mailing list, as including all of the
controllers and so forth on that one diagram will turn it into an unreadable mess in short order.

Likely the Class Diagram will be expanded to include packages and so forth.